+++ src/test/intellisense/Lib/AppGlobalScript.js
@@ -4,3 +4,5 @@
  */
 
 const USER = 'Maxim'
+
+const ROLE = 'Admin'