+++ src/test/intellisense/dep.js
@@ -9,3 +9,7 @@
 function sum(a, b) {
   return a + b
 }
+
+function multiply(a, b) {
+  return a * b
+}