+++ src/test/intellisense/entry.js
@@ -3,5 +3,5 @@ import 'dep.js'
 sum(1, 2)
 
 if (USER === 'Maxim') {
-  alert(IDOK)
+  unknownFunction(IDOK)
 }