+++ src/test/intellisense/formatting/region-in-region.js
@@ -1,0 +1,11 @@
+/**
+ * #text
+ * lorem
+ * #endtext
+ */
+
+var query = `
+#sql
+select 1 v
+#endsql
+`;